Rename json_* files to ovis_json_*
Ovis/LDMS's own custom json library name and header conflict with well
known existing json libraries. We change the names from json_* to
ovis_json_* to avoid conflicts.

Fixes #222 for v4
