+++ ldms/src/ldmsd/test/ldmsd_stream_publish.c
@@ -9,7 +9,7 @@
 #include <getopt.h>
 #include <semaphore.h>
 #include <pthread.h>
-#include <json/json_util.h>
+#include <ovis_json/ovis_json_util.h>
 #include <ovis_util/util.h>
 #include "ldms.h"
 #include "../ldmsd_request.h"
