+++ ldms/src/sampler/papi/papi_sampler.c
@@ -55,7 +55,7 @@
 #include <unistd.h>
 #include <assert.h>
 #include <papi.h>
-#include <json/json_util.h>
+#include <ovis_json/ovis_json_util.h>
 #include "ldms.h"
 #include "ldmsd.h"
 #include "ldmsd_stream.h"
